refactor ExceptionStackLine to seperate file in database layer
